improve ArtCompiler performance
Once the node have been written to the file, it is not
accessed anymore, so all modification can directly occurs
in RAM.

Also, all optimisations are done once a trie path have been replaced
by a new one, so optimisations is done locally just before node
are inserted.

As a new path only happen when there is a path split, path compression
can't optimise this part so it is not problematic that optimisations
are not done globally (other optimisations are per nodes).

The performance improvement require that all word are sorted before they are added